FPGA SPI: Complete rewrite, don't use much of at91lib
The problem is that the SPI routines of at91lib will cause CS to go
high between the individual bytes of a transfer, causing the state
machine in the FPGA to re-start (during write).

CSAAT doesn't seem to work in the hardware, as we seem to loose TDRE
events after transmit in this configuration.

The current code is fast enough to make sure CS doesn't get inactive
during the transfer of 5 octets.